+++ driver/bluetooth_low_energy.c
@@ -87,7 +87,7 @@ uint8_t ring_pop(ring_buf_t *ring)
 /**
  * Send a buffer out, retrying continuously until it goes to completion (with success or failure).
  */
-static void ble_tx(ble_service_t *service, uint8_t const *buf, uint16_t len)
+void ble_tx(ble_service_t *service, uint8_t const *buf, uint16_t len)
 {
     nrfx_err_t err;
     ble_gatts_hvx_params_t hvx_params = {
@@ -113,71 +113,4 @@ static void ble_tx(ble_service_t *service, uint8_t const *buf, uint16_t len)
 
     // Catch other errors
     app_err(err);
-}
-
-void ble_raw_tx(uint8_t const *buf, uint16_t len)
-{
-    ble_tx(&ble_raw_service, buf, len);
-}
-
-/**
- * Sends all buffered data in the tx ring buffer over BLE.
- */
-static void ble_nus_flush_tx(void)
-{
-    // Local buffer for sending data
-    uint8_t buf[BLE_MAX_MTU_LENGTH] = "";
-    uint16_t len = 0;
-
-    // If not connected, do not flush.
-    if (ble_conn_handle == BLE_CONN_HANDLE_INVALID)
-        return;
-
-    // If there's no data to send, simply return
-    if (ring_empty(&nus_tx))
-        return;
-
-    // For all the remaining characters, i.e until the heads come back together
-    while (!ring_empty(&nus_tx))
-    {
-        // Copy over a character from the tail to the outgoing buffer
-        buf[len++] = ring_pop(&nus_tx);
-
-        // Break if we over-run the negotiated MTU size, send the rest later
-        if (len >= ble_negotiated_mtu)
-            break;
-    }
-    ble_tx(&ble_nus_service, buf, len);
-}
-
-int ble_nus_rx(void)
-{
-    while (ring_empty(&nus_rx))
-    {
-        // While waiting for incoming data, we can push outgoing data
-        ble_nus_flush_tx();
-
-        // If there's nothing to do
-        if (ring_empty(&nus_tx) && ring_empty(&nus_rx))
-            // Wait for events to save power
-            sd_app_evt_wait();
-    }
-
-    // Return next character from the RX buffer.
-    return ring_pop(&nus_rx);
-}
-
-void ble_nus_tx(char const *buf, size_t len)
-{
-    for (size_t i = 0; i < len; i++)
-    {
-        while (ring_full(&nus_tx))
-            ble_nus_flush_tx();
-        ring_push(&nus_tx, buf[i]);
-    }
-}
-
-bool ble_nus_is_rx_pending(void)
-{
-    return ring_empty(&nus_rx);
 }
