+++ mphalport.c
@@ -30,6 +30,7 @@
 #include "driver/bluetooth_low_energy.h"
 #include "nrfx_timer.h"
 #include "nrfx_systick.h"
+#include "nrf_soc.h"
 
 const char help_text[] = {
     "Welcome to MicroPython!\n\n"
@@ -105,14 +106,61 @@ mp_import_stat_t mp_import_stat(const char *path)
     return MP_IMPORT_STAT_NO_EXIST;
 }
 
+/**
+ * Sends all buffered data in the tx ring buffer over BLE.
+ */
+static void ble_nus_flush_tx(void)
+{
+    // Local buffer for sending data
+    uint8_t buf[BLE_MAX_MTU_LENGTH] = "";
+    uint16_t len = 0;
+
+    // If not connected, do not flush.
+    if (ble_conn_handle == BLE_CONN_HANDLE_INVALID)
+        return;
+
+    // If there's no data to send, simply return
+    if (ring_empty(&nus_tx))
+        return;
+
+    // For all the remaining characters, i.e until the heads come back together
+    while (!ring_empty(&nus_tx))
+    {
+        // Copy over a character from the tail to the outgoing buffer
+        buf[len++] = ring_pop(&nus_tx);
+
+        // Break if we over-run the negotiated MTU size, send the rest later
+        if (len >= ble_negotiated_mtu)
+            break;
+    }
+    ble_tx(&ble_nus_service, buf, len);
+}
+
 int mp_hal_stdin_rx_chr(void)
 {
-    return ble_nus_rx();
+    while (ring_empty(&nus_rx))
+    {
+        // While waiting for incoming data, we can push outgoing data
+        ble_nus_flush_tx();
+
+        // If there's nothing to do
+        if (ring_empty(&nus_tx) && ring_empty(&nus_rx))
+            // Wait for events to save power
+            sd_app_evt_wait();
+    }
+
+    // Return next character from the RX buffer.
+    return ring_pop(&nus_rx);
 }
 
 void mp_hal_stdout_tx_strn(const char *str, mp_uint_t len)
 {
-    ble_nus_tx(str, len);
+    for (size_t i = 0; i < len; i++)
+    {
+        while (ring_full(&nus_tx))
+            ble_nus_flush_tx();
+        ring_push(&nus_tx, str[i]);
+    }
 }
 
 void mp_hal_set_interrupt_char(char c)
